Add CloudOmeZarrServer skeleton
Adds a CloudOmeZarrServer and corresponding builder. For now, the
builder recognizes all files ending in .zattrs and applies no
validation. Then the image server hard-codes a 1000x1000 image with RGB
channels, filled with 100 tiles filled with color 0xff00ff (pink).

When added to QuPath, this renders a loaded .zattrs file as a grid of
pink tiles. Yay :)