Don't assume forward slash before gs:/
On Windows, it looks like we may actually get a backslash before gs:/.
This checks out as it seems to be using the gs://bucket/file format as a
relative path, turning it into:

```
file:/C:\working-directory\gs:/bucket/file.
```

So let's just test for gs:/ not whatever character comes in before.

All this is due to this code (I think) :

https://github.com/qupath/qupath/blob/4f53be759c1858f92aeb396741a1fc60161bbe3f/qupath-core/src/main/java/qupath/lib/images/servers/ImageServerProvider.java#L140

Since a gs:// url doesn't start with file or http, the code treats it as
a regular path.

Let's hope this fixes it …