Read tile channels based on active channels
To reduce considerable wasted reads on channels that we aren't even
displaying, the Cloud OME Zarr Server now only reads channel data that
is currently active according to the imageDisplay settings.

Due to QuPath caching upstream (in the region store, not just the image
server itself) we clear the entire cache whenever the imageDisplay
changes. This forces QuPath to re-build its tile cache, this time
fetching the new set of selected channels.