Add 'package' step to publish to GH Packages
To facilitate downloading the extension code for use elsewhere, let's
publish to GitHub packages during build.

Supposedly we should be able to reuse the 0.1.0-snapshot version, but if
the snapshot mechanism isn't supported we'll need to keep incrementing
the version.

Also, update the build machine from Java 17 to Java 21 to align with IDE
build settings.